+++ src/math/vec2.rs
@@ -1,7 +1,7 @@
 #[derive(Debug)]
 pub struct Vec2 {
-    pub x: f64,
-    pub y: f64,
+    pub x: f32,
+    pub y: f32,
 }
 
 impl Vec2 {
@@ -17,19 +17,19 @@ impl Vec2 {
         }
     }
 
-    pub fn scale(&mut self, scale: f64) {
+    pub fn scale(&mut self, scale: f32) {
         self.x *= scale;
         self.y *= scale;
     }
 
-    pub fn length(&self) -> f64 {
-        f64::sqrt(self.x * self.x + self.y * self.y)
+    pub fn length(&self) -> f32 {
+        f32::sqrt(self.x * self.x + self.y * self.y)
     }
 
-    pub fn distance(&self, v0: &Vec2) -> f64 {
+    pub fn distance(&self, v0: &Vec2) -> f32 {
         let x = self.x - v0.x;
         let y = self.y - v0.y;
-        f64::sqrt(x * x + y * y)
+        f32::sqrt(x * x + y * y)
     }
 }
